Ejercicio 4 Terminado (Por fín)

+++ Room/Ejercicio4/Practica4/app/src/main/java/com/example/practica4/Dao/FavoritoDao.kt
@@ -10,7 +10,7 @@ import com.example.practica4.Entity.FavoritoEntity
 @Dao
 interface FavoritoDao {
     @Query("SELECT * FROM FavoritoEntity WHERE usuarioId = :userId")
-    suspend fun obtenerTodosLosFavoritos(userId: Int): MutableList<FavoritoEntity>
+    suspend fun obtenerTodosLosFavoritos(userId: Long): MutableList<FavoritoEntity>
     @Insert
     suspend fun agregarFavorito(favoritoEntity: FavoritoEntity)
     @Delete
